Rename package to filtermetric
