+++ internal/processor/filtermetric/doc.go
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package metric is a helper package for processing metrics.
-package metric
+// Package filtermetric is a helper package for processing metrics.
+package filtermetric
